requirements.txt
pandas>=2.0.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
# LangChain dependencies
langchain>=0.0.245
langchain-openai>=0.0.2
langchain-core>=0.0.14
langchain-community>=0.0.8
pydantic>=2.0.0
# Concurrent processing
joblib>=1.3.0
